Explicitly store 32-bit values in workmem

+++ lz4_ssparse.h
@@ -32,7 +32,7 @@ static size_t
 lz4_ssparse_workmem_size(size_t src_size)
 {
 	return (LOOKUP_SIZE < 2 * src_size ? 3 * src_size : src_size + LOOKUP_SIZE)
-	     * sizeof(unsigned long);
+	     * sizeof(uint32_t);
 }
 
 static unsigned long
@@ -69,11 +69,11 @@ lz4_pack_ssparse(const void *src, void *dst, unsigned long src_size, void *workm
 	// One detail is that we actually use src_size + 1 elements of cost,
 	// but we put mpos after it, where we do not need the first element.
 	//
-	unsigned long *const prev = (unsigned long *) workmem;
-	unsigned long *const mpos = prev + src_size;
-	unsigned long *const mlen = mpos + src_size;
-	unsigned long *const cost = prev;
-	unsigned long *const lookup = mpos;
+	uint32_t *const prev = (uint32_t *) workmem;
+	uint32_t *const mpos = prev + src_size;
+	uint32_t *const mlen = mpos + src_size;
+	uint32_t *const cost = prev;
+	uint32_t *const lookup = mpos;
 
 	// Phase 1: Build hash chains
 	const int bits = 2 * src_size < LOOKUP_SIZE ? LZ4_HASH_BITS : lz4_log2(src_size);
@@ -157,13 +157,13 @@ lz4_pack_ssparse(const void *src, void *dst, unsigned long src_size, void *workm
 			// max length will always be longer or equal, so we need
 			// only consider the extension.
 			if (len > max_len) {
-				unsigned long min_cost = ULONG_MAX;
+				unsigned long min_cost = UINT32_MAX;
 				unsigned long min_cost_len = 3;
 
 				// Find lowest cost match length
 				for (unsigned long i = max_len + 1; i <= len; ++i) {
 					unsigned long match_cost = lz4_match_cost(i);
-					assert(match_cost < ULONG_MAX - cost[cur + i]);
+					assert(match_cost < UINT32_MAX - cost[cur + i]);
 					unsigned long cost_here = match_cost + cost[cur + i];
 
 					if (cost_here < min_cost) {
